Found that because the time variables were an int, it reached the peak value after about 4 consecutive runs. Changed time from int to unsigned long. As a result, the music box can work for a further duration.

+++ Complete_Test_With_LEDs/Complete_Test_With_LEDs.ino
@@ -33,10 +33,10 @@ int songLength = 14;
 // OTHER IMPORTANT VARIABLES //
 boolean solenoidFound = false;
 boolean readyForSong = true;
-int initTime = 0;
+unsigned long initTime = 0;
 int initDelay = 500;
-int nextNoteTime = 0;
-int solenoidDropTime = 0;
+unsigned long nextNoteTime = 0;
+unsigned long                                                                                                                                                                                                                                                                       solenoidDropTime = 0;
 int holdTime = 8000;
